Remove AbstractMIDletSuiteLoader class
